renamed some components for better understanding of the purpose
