+++ src/components/AppLayout/HomeView/ConversationView/ConversationView.tsx
@@ -1,23 +1,23 @@
 import MessageInput from './MessageInput/MessageInput';
 import MessagesList from './MessagesList/MessagesList';
-import styles from './ConvoView.module.css';
+import styles from './ConversationView.module.css';
 import { useSelector } from 'react-redux';
 import { RootState } from '../../../../types/appTypes';
-import ConvoHeader from './ConvoHeader/ConvoHeader';
-import useConvoSub from '../../../../hooks/useConvoSub';
+import useConversationSub from '../../../../hooks/useConversationSub';
 import { ThreeCircles } from 'react-loader-spinner';
+import ConversationHeader from './ConversationHeader/ConversationHeader';
 
-function ConvoView() {
+function ConversationView() {
   const currentConvoData = useSelector(
     (state: RootState) => state.home.current_convo
   );
 
-  const isLoadingMessages = useConvoSub(currentConvoData?.channel.id);
+  const isLoadingMessages = useConversationSub(currentConvoData?.channel.id);
 
   return (
     <div className={styles['messages-view']}>
       <div>
-        <ConvoHeader
+        <ConversationHeader
           type={currentConvoData.channel.type}
           name={currentConvoData.channel.name}
         />
@@ -41,4 +41,4 @@ function ConvoView() {
   );
 }
 
-export default ConvoView;
+export default ConversationView;
